src/predictor.py
import pandas as pd
import joblib
from rdkit import Chem
from rdkit.Chem import AllChem, Descriptors
import os

def generate_features(smiles_list, descriptor_type):
    """Generates the specific features required by the recommended model."""
    mols = [Chem.MolFromSmiles(s) for s in smiles_list]
    
    if descriptor_type == "ECFP Fingerprints (1024-bit)":
        fps = [AllChem.GetMorganFingerprintAsBitVect(m, 2, nBits=1024) for m in mols]
        return fps
    
    elif descriptor_type == "RDKit Physicochemical Descriptors":
        # Genera las 22 propiedades que mencionaste en tu reporte
        calc = []
        for m in mols:
            props = [Descriptors.MolLogP(m), Descriptors.MolWt(m), Descriptors.NumHDonors(m)] # Simplificado
            calc.append(props)
        return calc
    
    return None

def run_prediction(df, model_name, descriptor_type):
    """
    Loads the pre-trained model and returns predictions.
    Note: You should place your .pkl files in the 'models/' folder.
    """
    # Placeholder: En producción aquí cargarías tu modelo real:
    # model = joblib.load(f'models/{model_name}.pkl')
    
    features = generate_features(df['smiles'], descriptor_type)
    
    # Simulación de predicción (Probabilidades)
    # y_pred = model.predict_proba(features)[:, 1]
    import numpy as np
    df['Probability'] = np.random.uniform(0, 1, len(df))
    df['Prediction'] = df['Probability'].apply(lambda x: 'Active' if x > 0.5 else 'Inactive')
    
    return df