requirements.txt
# --- Core Framework ---
streamlit>=1.25.0

# --- Chemistry & Bioinformatics ---
# Nota: rdkit-pypi es la versión recomendada para entornos pip
rdkit-pypi>=2022.9.5

# --- Data Science & ML ---
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
scipy>=1.9.0
xgboost>=1.6.0

# --- Visualization ---
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.12.0